INS-751: safe math moved into wallet dir
